api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API for the SoftCon-2022 Library!"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Flask-SQLAlchemy = "^2.5.1"
flask-marshmallow = "^0.14.0"
connexion = {extras=["swagger-ui"], version = "^2.10.0"}
marshmallow-sqlalchemy = "^0.27.0"
Flask-Cors = "^3.0.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
